ReverseGeocode

ReverseGeocode.COLUMNS

ReverseGeocode.Geocoder

ReverseGeocode.decode

ReverseGeocode.decode

ReverseGeocode.download_country_codes

ReverseGeocode.download_data

ReverseGeocode.read_data

ReverseGeocode.COLUMNS — Constant

Column names in the geonames dumpfile (from http://download.geonames.org/export/dump):

ReverseGeocode.Geocoder — Type

Geocoder(;data_dir="./data", geo_file="cities1000"))

Geocoder structure that holds the reference points and their labels.

ReverseGeocode.decode — Method

decode(gc::Geocoder, point) => NamedTuple{(:country, :country_code, :city), Tuple{String, String, String}}

Decode for single point. If processing many points, it should be faster with decode(gc, points) than with this function.

ReverseGeocode.decode — Method

decode(gc::Geocoder, points) => Array{NamedTuple{(:country, :country_code, :city), Tuple{String, String, String}}}

Return country name and city for collection of points. Points should be either an array of staticaly sized arrays (e.g. StaticArrays) or a Matrix (see NearestNeighbors.jl documentation for details).

The country and city is determined by the nearest neighbor search in the labelled list of city locations from geonames.org. As such, the results may not be exactly accurate (e.g searches for points close to borders or in the middle of nowhere).

Nearest neighbor search uses the euclidian metric in the space of lat/lon coordinates.

# Example
julia> gc = Geocoder();
julia> ReverseGeocode.decode(gc, [SA[49.5863897, 17.2627342], SA[63.3342550, 12.0280064]])
2-element Array{Tuple{String,String},1}:
 (country="Czechia", country_code="CZ", city="Olomouc")
 (country="Norway", country_code="NO", city="Meråker")

ReverseGeocode.download_country_codes — Method

Download and resave the country codes csv from geonames.

ReverseGeocode.download_data — Method

download_data(;data_dir="./data", geo_file="cities1000", header=COLUMNS)

Download dump from geonames.org. This function fetches a file of all cities with a population > 1000 (other options are 500,5000,15000) (or seats of admin div). The dump is unpacked and city name, coordinates and country code are resaved in a csv file for use in the Geocoder.

ReverseGeocode.read_data — Method

read_data(;data_dir="./data", geo_file="cities1000")

Load coordinates, countrycodes and city names from the csv export of the geonames file. Make sure to call downloaddata() before read_data().